+++ src/lib/tta_enc.c
@@ -89,7 +89,7 @@ tta_encode_2ch(
 
 //////////////////////////////////////////////////////////////////////////////
 
-// returns LIBTTAr_RET_OK (0) on success
+// returns LIBTTAr_RET_DONE or LIBTTAr_RET_AGAIN on success
 int
 libttaR_tta_encode(
 	u8 *const dest, const i32 *const src,
@@ -103,7 +103,7 @@ libttaR_tta_encode(
 		*user
 @*/
 {
-	size_t r = LIBTTAr_RET_AGAIN;
+	int r = LIBTTAr_RET_AGAIN;
 	size_t nbytes_tta_enc;
 	const  u8 predict_k    = tta_predict_k(samplebytes);
 	const i32 filter_round = tta_filter_round(samplebytes);
